refactor modal add, edit & delete on sysconfig

+++ src/views/system-configuration/approval/ListApproval.vue
@@ -8,11 +8,15 @@
     import { ref, computed, onBeforeMount } from 'vue'
     import arrowicon from "@/assets/navbar/icon_arrow.svg";
     import ModalEditApproval from '@/components/system-configuration/approval/ModalEditApprover.vue'
-    import ModalDelete from '@/components/modal/modalDelete.vue'
 
+    import Swal from "sweetalert2"
     import Api from '@/utils/Api'
 
     import tableContainer from '@/components/table/tableContainer.vue'
+
+    import deleteicon from "@/assets/navbar/delete_icon.svg";
+    import icondanger from "@/assets/Danger.png";
+    import iconClose from "@/assets/navbar/icon_close.svg";
 
     import { useFormAddStore } from '@/stores/sysconfig/add-modal.js'
     import { useFormEditStore } from '@/stores/sysconfig/edit-modal.js'
@@ -137,14 +141,64 @@
     }
 
     const deleteData = (event) => {
-      console.log(event)
-      Api.delete(`/approval/delete_data_approval/${event}`)
-      fetch()
-      if (sortedData.value.length == 1) {
-        router.go()
-      } else {
-        fetch()
-      }
+
+      Swal.fire({
+        title:
+          "<span class='font-JakartaSans font-medium text-[28px]'>Are you sure want to delete this?</span>",
+        html: "<div class='font-JakartaSans font-medium text-sm'>This will delete this data permanently, You cannot undo this action.</div>",
+        iconHtml: `<img src="${icondanger}" />`,
+        showCloseButton: true,
+        closeButtonHtml: `<img src="${iconClose}" class="hover:scale-75"/>`,
+        showCancelButton: true,
+        buttonsStyling: false,
+        cancelButtonText: "Cancel",
+        customClass: {
+          cancelButton: "swal-cancel-button",
+          confirmButton: "swal-confirm-button",
+        },
+        reverseButtons: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes",
+      })
+        
+      .then((result) => {
+        if (result.isConfirmed) {
+          Api.delete(`/approval/delete_data_approval/${event}`)
+          .then((res) => {
+
+            Swal.fire({
+              title: "Successfully",
+              text: "Data has been deleted.",
+              icon: "success",
+              showCancelButton: false,
+              confirmButtonColor: "#015289",
+              showConfirmButton: false,
+              timer: 1500,
+            });
+
+            if (sortedData.value.length == 1) {
+              // router.go() 
+              fetch()
+            } else {
+              fetch()
+            }
+
+          })
+        } else {
+          return
+        }
+      })
+
+
+      // Api.delete(`/approval/delete_data_approval/${event}`)
+      // fetch()
+      // if (sortedData.value.length == 1) {
+      //   router.go()
+      // } else {
+      //   fetch()
+      // }
+
     }
 
     const getData = () => {
@@ -213,8 +267,14 @@
                       {{ data.menu }}
                     </td>
                     <td class="flex flex-wrap gap-4 justify-center">
-                      <ModalEditApproval @fetchApproval="getData" @edit-approver="editExistingApprover(data.id)" :formContent="[data.approval_name, data.id_company, data.id_menu, data.id_code_document, data.detail, data.detail[index].id_matrix]" />
-                      <ModalDelete @confirm-delete="deleteData(data.id)" />
+                      <ModalEditApproval @fetchApproval="getData" @edit-approver="editExistingApprover(data.id)" :formContent="[data.approval_name, data.id_company, data.id_menu, data.id_code_document, data.detail]" />
+                      <!-- approval authorities nya gaboleh kosong, karena akan dikirim sebagai props -->
+                      <!-- {{ index }} -->
+                     <!-- {{ data.detail[index] }} -->
+                      <!-- <ModalDelete @confirm-delete="deleteData(data.id)" /> -->
+                      <button @click="deleteData(data.id)">
+                        <img :src="deleteicon" class="w-6 h-6" />
+                      </button>
                     </td>
                   </tr>
