+++ src/views/system-configuration/menu/ListMenu.vue
@@ -4,6 +4,7 @@
     import TableTopBar from '@/components/layout/TableTopBar.vue'
     import Footer from '@/components/layout/Footer.vue'
 
+    import Swal from "sweetalert2"
     import Api from '@/utils/Api'
     import router from '@/router'
     
@@ -12,9 +13,12 @@
     import arrowicon from "@/assets/navbar/icon_arrow.svg"
 
     import ModalEditMenu from '@/components/system-configuration/menu/ModalEditMenu.vue'
-    import ModalDelete from '@/components/modal/modalDelete.vue'
-
+    
     import tableContainer from '@/components/table/tableContainer.vue'
+
+    import deleteicon from "@/assets/navbar/delete_icon.svg";
+    import icondanger from "@/assets/Danger.png";
+    import iconClose from "@/assets/navbar/icon_close.svg";
 
     import { useSidebarStore } from "@/stores/sidebar.js"
     import { useFormEditStore } from '@/stores/sysconfig/edit-modal.js'
@@ -44,12 +48,62 @@
     }
 
     const deleteData = async (event) => {
-     await Api.delete(`/menu/delete_data/${event}`)
-      if (sortedData.value.length == 1) {
-        router.go()
-      } else {
-        fetch()
-      }
+
+      Swal.fire({
+        title:
+          "<span class='font-JakartaSans font-medium text-[28px]'>Are you sure want to delete this?</span>",
+        html: "<div class='font-JakartaSans font-medium text-sm'>This will delete this data permanently, You cannot undo this action.</div>",
+        iconHtml: `<img src="${icondanger}" />`,
+        showCloseButton: true,
+        closeButtonHtml: `<img src="${iconClose}" class="hover:scale-75"/>`,
+        showCancelButton: true,
+        buttonsStyling: false,
+        cancelButtonText: "Cancel",
+        customClass: {
+          cancelButton: "swal-cancel-button",
+          confirmButton: "swal-confirm-button",
+        },
+        reverseButtons: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes",
+      })
+        
+      .then((result) => {
+        if (result.isConfirmed) {
+          Api.delete(`/menu/delete_data/${event}`)
+          .then((res) => {
+
+            Swal.fire({
+              title: "Successfully",
+              text: "Data has been deleted.",
+              icon: "success",
+              showCancelButton: false,
+              confirmButtonColor: "#015289",
+              showConfirmButton: false,
+              timer: 1500,
+            });
+
+            if (sortedData.value.length == 1) {
+              // router.go()
+              fetch()
+            } else {
+              fetch()
+            }
+
+          })
+        } else {
+          return
+        }
+      })
+
+    //  await Api.delete(`/menu/delete_data/${event}`)
+    //   if (sortedData.value.length == 1) {
+    //     router.go()
+    //   } else {
+    //     fetch()
+    //   }
+
     }
 
     const addNewMenu = async () => {
@@ -141,8 +195,6 @@
 
     let status = ref('')
     let message = ref('')
-
-    // watch(ref, callback)
 
     const fetch = async () => {
       try {
@@ -255,7 +307,9 @@
                         <td class="flex flex-wrap justify-center h-full gap-4 relative">
                           <div class="flex items-center absolute top-0 bottom-0">
                             <ModalEditMenu @unlock-scrollbar="lockScrollbar = !lockScrollbar" @change-menu="editMenu(data.id)" :formContent="[data.menu, data.url, data.sort, data.icon]" />
-                            <ModalDelete @confirm-delete="deleteData(data.id)" />
+                            <button @click="deleteData(data.id)">
+                              <img :src="deleteicon" class="w-6 h-6" />
+                            </button>
                           </div>
                         </td>
