+++ src/views/system-configuration/sequence/ListSequence.vue
@@ -4,15 +4,19 @@
     import TableTopBar from '@/components/layout/TableTopBar.vue'
     import Footer from '@/components/layout/Footer.vue'
 
+    import Swal from "sweetalert2"
     import Api from '@/utils/Api'
 
     // import untuk user table
     import { ref, computed, onBeforeMount } from 'vue'
     import arrowicon from "@/assets/navbar/icon_arrow.svg"
     import ModalEditSequence from '@/components/system-configuration/sequence/ModalEditSequence.vue'
-    import ModalDelete from '@/components/modal/modalDelete.vue'
 
     import tableContainer from '@/components/table/tableContainer.vue'
+
+    import deleteicon from "@/assets/navbar/delete_icon.svg";
+    import icondanger from "@/assets/Danger.png";
+    import iconClose from "@/assets/navbar/icon_close.svg";
 
     import { useFormEditStore } from '@/stores/sysconfig/edit-modal.js'
     import { useFormAddStore } from '@/stores/sysconfig/add-modal.js'
@@ -112,12 +116,54 @@
     }
 
     const deleteData = async (event) => {
-     await Api.delete(`/sequence/delete_data/${event}`)
-      if (sortedData.value.length == 1) {
-        router.go()
-      } else {
-        fetch()
-      }
+
+      Swal.fire({
+        title:
+          "<span class='font-JakartaSans font-medium text-[28px]'>Are you sure want to delete this?</span>",
+        html: "<div class='font-JakartaSans font-medium text-sm'>This will delete this data permanently, You cannot undo this action.</div>",
+        iconHtml: `<img src="${icondanger}" />`,
+        showCloseButton: true,
+        closeButtonHtml: `<img src="${iconClose}" class="hover:scale-75"/>`,
+        showCancelButton: true,
+        buttonsStyling: false,
+        cancelButtonText: "Cancel",
+        customClass: {
+          cancelButton: "swal-cancel-button",
+          confirmButton: "swal-confirm-button",
+        },
+        reverseButtons: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes",
+      })
+        
+      .then((result) => {
+        if (result.isConfirmed) {
+          Api.delete(`/sequence/delete_data/${event}`)
+          .then((res) => {
+
+            Swal.fire({
+              title: "Successfully",
+              text: "Data has been deleted.",
+              icon: "success",
+              showCancelButton: false,
+              confirmButtonColor: "#015289",
+              showConfirmButton: false,
+              timer: 1500,
+            });
+
+            if (sortedData.value.length == 1) {
+              router.go()
+            } else {
+              fetch()
+            }
+
+          })
+        } else {
+          return
+        }
+      })
+
     }
 
     const addNewSequence = () => {
@@ -237,7 +283,9 @@
                       </td>
                       <td class="flex flex-wrap gap-4 justify-center">
                         <ModalEditSequence @change-sequence="editExistingSequence(data.id)" :formContent="[data.sequence_name, data.prefix, data.suffix, data.sequence_size, data.recycle, data.next_value, data.id_menu, data.id_company]" />
-                        <ModalDelete @confirm-delete="deleteData(data.id)" />
+                        <button @click="deleteData(data.id)">
+                          <img :src="deleteicon" class="w-6 h-6" />
+                        </button>
                       </td>
                     </tr>
